Upgrade 8.0 (#6)
* Update release.yml
* Update go.mod
* update build.sh to follow migration to mono-repo
* Use mmctl 7.10.4 version
* call make dist instead of make build
* Update build script
* Upgrade GO_VERSION to 1.19

+++ build.sh
@@ -4,12 +4,13 @@ set -eux
 BUILD_USER_HOME="${BUILD_USER_HOME:-/build}"
 BUILD_USER_NAME="${BUILD_USER_NAME:-build}"
 # Debian release used during build
-DEBIAN_RELEASE="${DEBIAN_RELEASE:-stretch}"
+DEBIAN_RELEASE="${DEBIAN_RELEASE:-bookworm}"
 # Mattermost version to build
 MATTERMOST_RELEASE="${MATTERMOST_RELEASE:-v5.26.0}"
 MMCTL_RELEASE="${MMCTL_RELEASE:-v5.26.0}"
+MM_FOCALBOARD_RELEASE="${MM_FOCALBOARD_RELEASE:-v5.26.0}"
 # golang version
-GO_VERSION="${GO_VERSION:-1.18.1}"
+GO_VERSION="${GO_VERSION:-1.19.9}"
 
 if [ "$(id -u)" -eq 0 ]; then # as root user
 	# create build user, if needed
@@ -41,21 +42,28 @@ if [ "$(id -u)" -eq 0 ]; then # as root user
 	apt-get update
 	# install dependencies
 	apt-get install --quiet \
-		wget build-essential patch git python2 
+		wget build-essential patch git python3 
 	# install 'pngquant' build dependencies (required by node module)
 	apt-get build-dep --quiet \
 		pngquant
 	# install go from golang.org
-	wget https://golang.org/dl/go${GO_VERSION}.linux-amd64.tar.gz
-	tar -xvf go${GO_VERSION}.linux-amd64.tar.gz
-	mv go /usr/local
+	if [ ! -f go${GO_VERSION}.linux-amd64.tar.gz ]; then
+		wget https://golang.org/dl/go${GO_VERSION}.linux-amd64.tar.gz
+	fi
+	if [ ! -d /usr/local/go ]; then
+		tar -xvf go${GO_VERSION}.linux-amd64.tar.gz
+		mv go /usr/local
+	fi
 	export GOROOT=/usr/local/go
 	export PATH=$GOROOT/bin:$PATH
 	# FIXME go (executed by build user) writes to GOROOT
 	install --directory --owner="${BUILD_USER_NAME}" \
 		"$(go env GOROOT)/pkg/$(go env GOOS)_$(go env GOARCH)"
+
+	echo "GOOS_GOARCH=$(go env GOOS)_$(go env GOARCH)"
 	# Re-invoke this build.sh script with the 'build' user
 	runuser -u "${BUILD_USER_NAME}" -- "${0}"
+
 	# salvage build artifacts
 	cp --verbose \
 		"${BUILD_USER_HOME}/mattermost-${MATTERMOST_RELEASE}-$(go env GOOS)-$(go env GOARCH).tar.gz" \
@@ -74,72 +82,58 @@ export NVM_DIR="$HOME/.nvm"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
 
 # download and extract Mattermost sources
-for COMPONENT in server webapp; do
-	install --directory "${HOME}/go/src/github.com/mattermost/mattermost-${COMPONENT}"
-	wget --quiet --continue --output-document="mattermost-${COMPONENT}.tar.gz" \
-		"https://github.com/mattermost/mattermost-${COMPONENT}/archive/${MATTERMOST_RELEASE}.tar.gz"
-	tar --directory="${HOME}/go/src/github.com/mattermost/mattermost-${COMPONENT}" \
-		--strip-components=1 --extract --file="mattermost-${COMPONENT}.tar.gz"
-done
+install --directory "${HOME}/go/src/github.com/mattermost/mattermost"
+wget --quiet --continue --output-document="mattermost.tar.gz" \
+	"https://github.com/mattermost/mattermost/archive/${MATTERMOST_RELEASE}.tar.gz"
+tar --directory="${HOME}/go/src/github.com/mattermost/mattermost" \
+	--strip-components=1 --extract --file="mattermost.tar.gz"
+
 # download and extract focalboard
 install --directory "${HOME}/go/src/github.com/mattermost/focalboard"
-	wget --quiet --continue --output-document="focalboard.tar.gz" \
-		"https://github.com/mattermost/focalboard/archive/${MATTERMOST_RELEASE}.tar.gz"
-	tar --directory="${HOME}/go/src/github.com/mattermost/focalboard" \
-		--strip-components=1 --extract --file="focalboard.tar.gz"
+wget --quiet --continue --output-document="focalboard.tar.gz" \
+	"https://github.com/mattermost/focalboard/archive/${MM_FOCALBOARD_RELEASE}.tar.gz"
+tar --directory="${HOME}/go/src/github.com/mattermost/focalboard" \
+	--strip-components=1 --extract --file="focalboard.tar.gz"
 
 # install mattermost-webapp's required version of nodejs
-pushd "${HOME}/go/src/github.com/mattermost/mattermost-webapp"
+pushd "${HOME}/go/src/github.com/mattermost/mattermost/webapp"
 nvm install
 popd
 
 # prepare the go build environment
 install --directory "${HOME}/go/bin"
-if [ "$(go env GOOS)_$(go env GOARCH)" != 'linux_amd64' ]; then
+if [ "$(go env GOOS)_$(go env GOARCH)" != 'linux_amd64' ] && [ ! -f "${HOME}/go/bin/linux_amd64" ] && [ ! -h "${HOME}/go/bin/linux_amd64" ]; then
 	ln --symbolic \
 		"${HOME}/go/bin/$(go env GOOS)_$(go env GOARCH)" \
 		"${HOME}/go/bin/linux_amd64"
 fi
-# build mmctl
-install --directory "${HOME}/go/src/github.com/mattermost/mmctl"
-wget --quiet --continue --output-document="mmctl.tar.gz" \
-	"https://github.com/mattermost/mmctl/archive/${MMCTL_RELEASE}.tar.gz"
-tar --directory="${HOME}/go/src/github.com/mattermost/mmctl" \
-	--strip-components=1 --extract --file="mmctl.tar.gz"
-find "${HOME}/go/src/github.com/mattermost/mmctl/" -type f -name '*.go' | xargs \
-	sed -i \
-	-e 's#//go:build linux || darwin#//go:build linux || darwin || dragonfly || freebsd || netbsd || openbsd#' \
-	-e 's#// +build linux darwin#// +build linux darwin dragonfly freebsd netbsd openbsd#'
-make --directory="${HOME}/go/src/github.com/mattermost/mmctl" \
-	BUILD_NUMBER="dev-$(go env GOOS)-$(go env GOARCH)-${MMCTL_RELEASE}" \
-	ADVANCED_VET=0 \
-	GO="GOARCH= GOOS= $(command -v go)"
+
 # build focalboard
 make --directory="${HOME}/go/src/github.com/mattermost/focalboard" \
 	prebuild
 make --directory="${HOME}/go/src/github.com/mattermost/focalboard" \
 	build
-# build Mattermost webapp
+# # build Mattermost webapp
 npm set progress false
 sed -i -e 's#--verbose#--display minimal#' \
-	"${HOME}/go/src/github.com/mattermost/mattermost-webapp/package.json"
-make --directory="${HOME}/go/src/github.com/mattermost/mattermost-webapp" \
-	build
+	"${HOME}/go/src/github.com/mattermost/mattermost/webapp/package.json"
+make --directory="${HOME}/go/src/github.com/mattermost/mattermost/webapp" \
+	dist
 # build Mattermost server
-patch --directory="${HOME}/go/src/github.com/mattermost/mattermost-server" \
-	--strip=1 -t < "${HOME}/build-release.patch"
+patch --directory="${HOME}/go/src/github.com/mattermost/mattermost/server" \
+	--strip=1 -t < "/root/build-release.patch"
 sed -i \
 	-e 's#go generate#env --unset=GOOS --unset=GOARCH go generate#' \
 	-e 's#$(GO) generate#env --unset=GOOS --unset=GOARCH go generate#' \
 	-e 's#PWD#CURDIR#' \
-	"${HOME}/go/src/github.com/mattermost/mattermost-server/Makefile" \
-	"${HOME}/go/src/github.com/mattermost/mattermost-server/build/release.mk"
-make --directory="${HOME}/go/src/github.com/mattermost/mattermost-server" \
+	"${HOME}/go/src/github.com/mattermost/mattermost/server/Makefile" \
+	"${HOME}/go/src/github.com/mattermost/mattermost/server/build/release.mk"
+make --directory="${HOME}/go/src/github.com/mattermost/mattermost/server" \
 	config-reset \
 	BUILD_NUMBER="dev-$(go env GOOS)-$(go env GOARCH)-${MATTERMOST_RELEASE}" \
 	GO="GOARCH= GOOS= $(command -v go)" \
 	PLUGIN_PACKAGES=''
-make --directory="${HOME}/go/src/github.com/mattermost/mattermost-server" \
+make --directory="${HOME}/go/src/github.com/mattermost/mattermost/server" \
 	build-linux package-linux \
 	BUILD_NUMBER="dev-$(go env GOOS)-$(go env GOARCH)-${MATTERMOST_RELEASE}" \
 	BUILD_BOARDS=0 \
@@ -147,7 +141,7 @@ make --directory="${HOME}/go/src/github.com/mattermost/mattermost-server" \
 	GO="GOARCH=$(go env GOARCH) GOOS=$(go env GOOS) $(command -v go)" \
 	PLUGIN_PACKAGES=''
 # rename archive and calculate its SHA512 sum
-mv "${HOME}/go/src/github.com/mattermost/mattermost-server/dist/mattermost-team-linux-amd64.tar.gz" \
+mv "${HOME}/go/src/github.com/mattermost/mattermost/server/dist/mattermost-team-linux-amd64.tar.gz" \
 	"${HOME}/mattermost-${MATTERMOST_RELEASE}-$(go env GOOS)-$(go env GOARCH).tar.gz"
 sha512sum "${HOME}/mattermost-${MATTERMOST_RELEASE}-$(go env GOOS)-$(go env GOARCH).tar.gz" | \
 	tee "${HOME}/mattermost-${MATTERMOST_RELEASE}-$(go env GOOS)-$(go env GOARCH).tar.gz.sha512sum"
